remove unused files and update jog comand

+++ ControlRoomApplication/ControlRoomApplication/Controllers/PLCCommunication/PLCDrivers/ScaleModelPLCDriver.cs
@@ -381,7 +381,7 @@ namespace ControlRoomApplication.Controllers {
             throw new NotImplementedException();
         }
 
-        public override bool Start_jog( RadioTelescopeAxisEnum axis , int speed , bool clockwise ) {
+        public override bool Start_jog( RadioTelescopeAxisEnum axis , double speed , bool clockwise ) {
             throw new NotImplementedException();
 
         }
